create UserLimitsHelper + move logic to it
